requirements.txt
discord.py>=2.0.0
requests
aiohttp
python-dotenv
colorlog